feat: move ExtractCommandFromArgs function to utils.go and remove redundant file
